Moved View classes to their own declarations from inline
